+++ GestureAuthoringTools/LabanEditor/src/labanotation/tool/wavfilter.py
@@ -44,7 +44,7 @@ def calcFilter(raw, gauss):
         tempArray = (raw.transpose()).tolist()
         glist = gauss.tolist()
         for i in range(0, col):
-            temp = [tempArray[i][0]]*(len(glist)/2) + tempArray[i] + [tempArray[i][-1]]*(len(glist)/2)
+            temp = [tempArray[i][0]]*(int(len(glist)/2)) + tempArray[i] + [tempArray[i][-1]]*(int(len(glist)/2))
             for j in range(0, len(temp)-len(glist)+1):
                 tempOut.append(np.dot(temp[j:j+len(glist)],glist))
         filtered = np.array(tempOut)
